requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
Flask-Login==0.6.3
Flask-Caching==2.1.0
PyMySQL==1.1.0
cryptography==41.0.3
python-dotenv==1.0.0
openpyxl==3.1.5
DBUtils==3.1.2
redis==5.0.1
cachetools==5.3.1
pandas
requests==2.31.0
bcrypt
paramiko
flask-wtf
flask_cors
